extracted canvas logic out of the web worker due to cross-browser compatibility issues- fixed a couple bugs
