+++ src/components/game.css
@@ -2,15 +2,9 @@ canvas#room-canvas {
     border: 1px solid black;
     border-radius: 10px;
     box-shadow: 0px 0px 10px 1px white;
-<<<<<<< Updated upstream
-    height: auto;
-    width: 35%;
-    margin: 10px auto
-=======
     height: 400px;
     max-width: 100%;
     margin: 20px auto
->>>>>>> Stashed changes
 }
 
 .room-stuff {
